+++ CHANGELOG.md
@@ -3,3 +3,5 @@
 * added `Predicate.is_instance`
 * added `filter_out_false` and `filter_out_nones`
 * fixed typing for `DictObject`
+* added `SparseMatrix.shoot`
+* added a `SparseMatrix` constructor